+++ extensions/obsidian/src/components/CreateNoteForm.tsx
@@ -4,14 +4,15 @@ import NoteCreator from "../utils/NoteCreator";
 import { NoteFormPreferences, FormValue, Vault } from "../utils/interfaces";
 
 export function CreateNoteForm(props: { vault: Vault }) {
-  const vault = props.vault;
-  const pref: NoteFormPreferences = getPreferenceValues();
+  const { vault } = props;
   const { pop } = useNavigation();
 
-  function folders() {
-    const folderString = pref.folderActions;
-    if (folderString) {
-      const folders = folderString
+  const pref = getPreferenceValues<NoteFormPreferences>();
+  const { folderActions, tags, prefTag, prefPath } = pref;
+
+  function parseFolderActions() {
+    if (folderActions) {
+      const folders = folderActions
         .split(",")
         .filter((folder) => !!folder)
         .map((folder: string) => folder.trim());
@@ -20,39 +21,21 @@ export function CreateNoteForm(props: { vault: Vault }) {
     return [];
   }
 
-  function tags() {
-    const tagsString = pref.tags;
-    const prefTag = pref.prefTag;
-    if (!tagsString) {
+  function parseTags() {
+    if (!tags) {
       if (prefTag) {
         return [{ name: prefTag, key: prefTag }];
       }
       return [];
     }
-    const tags = tagsString
+    const parsedTags = tags
       .split(",")
       .map((tag) => ({ name: tag.trim(), key: tag.trim() }))
       .filter((tag) => !!tag);
     if (prefTag) {
-      tags.push({ name: prefTag, key: prefTag });
+      parsedTags.push({ name: prefTag, key: prefTag });
     }
-    return tags;
-  }
-
-  function prefTag(): string[] {
-    const prefTag = pref.prefTag;
-    if (prefTag) {
-      return [prefTag];
-    }
-    return [];
-  }
-
-  function prefPath(): string {
-    const prefPath = pref.prefPath;
-    if (prefPath) {
-      return prefPath;
-    }
-    return "";
+    return parsedTags;
   }
 
   async function createNewNote(noteProps: FormValue, path: string | undefined = undefined) {
@@ -60,10 +43,8 @@ export function CreateNoteForm(props: { vault: Vault }) {
       noteProps.path = path;
     }
     const nc = new NoteCreator(noteProps, vault, pref);
-    const saved = await nc.createNote();
-    if (saved) {
-      pop();
-    }
+    nc.createNote();
+    pop();
   }
 
   return (
@@ -72,7 +53,7 @@ export function CreateNoteForm(props: { vault: Vault }) {
       actions={
         <ActionPanel>
           <Action.SubmitForm title="Create" onSubmit={createNewNote} />
-          {folders()?.map((folder, index) => (
+          {parseFolderActions()?.map((folder, index) => (
             <Action.SubmitForm
               title={"Create in " + folder}
               onSubmit={(props: FormValue) => createNewNote(props, folder)}
@@ -83,14 +64,29 @@ export function CreateNoteForm(props: { vault: Vault }) {
         </ActionPanel>
       }
     >
-      <Form.TextField title="Name" id="name" placeholder="Name of note" />
-      <Form.TextField title="Path" id="path" defaultValue={prefPath()} placeholder="path/to/note (optional)" />
-      <Form.TagPicker id="tags" title="Tags" defaultValue={prefTag()}>
-        {tags()?.map((tag) => (
+      <Form.TextField
+        title="Name"
+        id="name"
+        placeholder="Name of note"
+        defaultValue={pref.fillFormWithDefaults ? pref.prefNoteName : ""}
+      />
+      <Form.TextField
+        title="Path"
+        id="path"
+        defaultValue={prefPath ? prefPath : ""}
+        placeholder="path/to/note (optional)"
+      />
+      <Form.TagPicker id="tags" title="Tags" defaultValue={prefTag ? [prefTag] : []}>
+        {parseTags()?.map((tag) => (
           <Form.TagPicker.Item value={tag.name.toLowerCase()} title={tag.name} key={tag.key} />
         ))}
       </Form.TagPicker>
-      <Form.TextArea title="Content:" id="content" placeholder={"Text"} />
+      <Form.TextArea
+        title="Content:"
+        id="content"
+        placeholder={"Text"}
+        defaultValue={pref.fillFormWithDefaults ? pref.prefNoteContent : ""}
+      />
     </Form>
   );
 }