+++ extensions/obsidian/src/searchNoteCommand.tsx
@@ -5,7 +5,7 @@ import { NoteListObsidian } from "./components/NoteListObsidian";
 import { VaultSelection } from "./components/VaultSelection";
 import { Vault } from "./utils/interfaces";
 import { NoVaultFoundMessage } from "./components/NoVaultFoundMessage";
-import { pathErrorToast } from "./components/Toasts";
+import { noVaultPathsToast } from "./components/Toasts";
 
 export default function Command() {
   const { ready, vaults } = useObsidianVaults();
@@ -19,6 +19,6 @@ export default function Command() {
   } else if (vaults.length == 1) {
     return <NoteListObsidian vault={vaults[0]} />;
   } else {
-    pathErrorToast();
+    noVaultPathsToast();
   }
 }