+++ extensions/obsidian/src/utils/NoteLoader.tsx
@@ -3,7 +3,7 @@ import fs from "fs";
 import path from "path";
 
 import { SearchNotePreferences, Note, Vault } from "./interfaces";
-import { getNoteFileContent } from "./utils";
+import { getNoteFileContent, tagsFor } from "./utils";
 
 class NoteLoader {
   vaultPath: string;
@@ -24,15 +24,17 @@ class NoteLoader {
       if (f_name) {
         name = f_name.split(".md")[0];
       }
-      const note = {
+
+      const noteContent = getNoteFileContent(f, false);
+
+      const note: Note = {
         title: name,
-        key: ++key,
         path: f,
-        content: getNoteFileContent(f),
+        tags: tagsFor(noteContent),
+        content: noteContent,
       };
       notes.push(note);
     }
-    // console.log("Loaded " + notes.length + " notes.");
     return notes;
   }
 
@@ -54,6 +56,7 @@ class NoteLoader {
         if (
           file.endsWith(".md") &&
           file !== ".md" &&
+          !file.includes(".excalidraw") &&
           !dirPath.includes(".obsidian") &&
           this.isValidFile(dirPath, exFolders)
         ) {